move ethereum client to relayer types
